refactor(view): move Stats to view layer and renamed to StatsView for clearer separation of concerns
