+++ src/main/kotlin/blackjack/controller/StatsManager.kt
@@ -2,17 +2,17 @@ package blackjack.controller
 
 import blackjack.model.Dealer
 import blackjack.model.Player
-import blackjack.model.Stats
+import blackjack.view.StatsView
 
 class StatsManager(players: List<Player>, dealer: Dealer) {
-    private var _winStatistics = Stats(players, dealer)
-    val winStatistics: Stats = _winStatistics
+    private var _winStatistics = StatsView(players, dealer)
+    val winStatistics: StatsView = _winStatistics
 
     fun processStatistics(
         players: List<Player>,
         dealer: Dealer,
     ) {
-        _winStatistics = Stats(players, dealer)
+        _winStatistics = StatsView(players, dealer)
         _winStatistics.updateDealerStats()
     }
 }
